pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt==3.2.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.41.0",
]
